+++ .storybook/main.js
@@ -1,0 +1,27 @@
+const path = require('path');
+
+module.exports = {
+  stories: ['../src/components/**/*.stories.tsx'],
+  addons: [
+    '@storybook/addon-knobs/register',
+    '@storybook/addon-actions/register',
+    '@storybook/addon-viewport/register',
+    '@storybook/addon-docs',
+  ],
+  presets: [
+    '@storybook/preset-scss',
+    {
+      name: '@storybook/preset-typescript',
+      options: {
+        tsLoaderOptions: {
+          configFile: path.resolve(__dirname, '../tsconfig.json'),
+          ignoreDiagnostics: [7005],
+        },
+        tsDocgenLoaderOptions: {
+          tsconfigPath: path.resolve(__dirname, '../tsconfig.json'),
+        },
+        include: [path.resolve(__dirname, '../')],
+      },
+    },
+  ],
+};