+++ src/components/FileSkeleton/FileSkeleton.stories.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 import FileSkeleton from '../FileSkeleton';
 import docs from './README.md';
 
-export default { title: 'FileSkeleton', parameters: { info: docs } };
+export default { title: 'FileSkeleton', component: FileSkeleton, parameters: { info: docs } };
 
 export const basic = () => <FileSkeleton />;