+++ src/components/ToolButton/ToolButton.stories.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import ToolButton from '../ToolButton';
 import docs from './README.md';
 
-export default { title: 'ToolButton', parameters: { info: docs } };
+export default { title: 'ToolButton', component: ToolButton, parameters: { info: docs } };
 
 export const basic = () => (
   <ToolButton disabled={boolean('disabled', false)} onClick={action('onClick')}>